todolist/todo_project/requirements.txt
ansible==8.7.0
ansible-core==2.15.8
apturl==0.5.2
asgiref==3.8.1
awscli==1.22.34
Babel==2.8.0
bcrypt==3.2.0
blinker==1.4
boto==2.49.0
boto3==1.34.23
botocore==1.34.23
Brlapi==0.8.3
certifi==2020.6.20
chardet==4.0.0
charset-normalizer==3.3.2
click==8.0.3
colorama==0.4.4
command-not-found==0.3
cowsay==6.1
cryptography==3.4.8
cupshelpers==1.0
dbus-python==1.2.18
defer==1.0.6
distro==1.7.0
distro-info==1.1+ubuntu0.1
Django==5.0.6
django-cors-headers==4.3.1
django-jazzmin==3.0.0
djangorestframework==3.15.1
docker==7.0.0
docutils==0.17.1
duplicity==0.8.21
fasteners==0.14.1
future==0.18.2
greenlet==1.1.2
gyp==0.1
httplib2==0.20.2
idna==3.3
importlib-metadata==4.6.4
jeepney==0.7.1
Jinja2==3.0.3
jmespath==0.10.0
keyring==23.5.0
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
louis==3.20.0
macaroonbakery==1.3.1
Mako==1.1.3
MarkupSafe==2.0.1
monotonic==1.6
more-itertools==8.10.0
msgpack==1.0.3
netifaces==0.11.0
ntlm-auth==1.4.0
oauthlib==3.2.0
olefile==0.46
packaging==21.3
paramiko==2.9.3
pexpect==4.8.0
Pillow==9.0.1
protobuf==3.12.4
ptyprocess==0.7.0
pyasn1==0.4.8
pycairo==1.20.1
pycups==2.0.1
Pygments==2.11.2
PyGObject==3.42.1
PyJWT==2.3.0
pykerberos==1.1.14
pymacaroons==0.13.0
PyNaCl==1.5.0
pynvim==0.4.2
pyparsing==2.4.7
PyQt5==5.15.10
PyQt5-Qt5==5.15.2
PyQt5-sip==12.13.0
pyRFC3339==1.1
pyTelegramBotAPI==4.16.1
python-apt==2.4.0+ubuntu2
python-dateutil==2.8.1
python-debian==0.1.43+ubuntu1.1
pytz==2022.1
pywinrm==0.3.0
pyxdg==0.27
PyYAML==5.4.1
reportlab==3.6.8
requests==2.31.0
requests-kerberos==0.12.0
requests-ntlm==1.1.0
resolvelib==0.8.1
roman==3.3
rsa==4.8
s3transfer==0.10.0
SecretStorage==3.3.1
six==1.16.0
sqlparse==0.4.2
ssh-import-id==5.11
systemd-python==234
typing_extensions==4.12.0
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36.1
unattended-upgrades==0.1
urllib3==1.26.5
usb-creator==0.3.7
wadllib==1.3.6
xdg==5
xkit==0.0.0
xmltodict==0.12.0
zipp==1.0.0
